feat(frontend): create shared router and entrypoints
