+++ Frontend/src/entry-client.tsx
@@ -1,9 +1,10 @@
-import { createRoot } from 'react-dom/client';
+import { hydrateRoot } from 'react-dom/client';
 import { createBrowserRouter } from 'react-router-dom';
+import routes, { AppRouter } from './router';
 import './index.css';
 import './i18n/index.ts';
-import routes, { AppRouter } from './router';
 
-createRoot(document.getElementById('root')!).render(
+hydrateRoot(
+    document.getElementById('root')!,
     <AppRouter router={createBrowserRouter(routes)} />
 );
